Revert "Added some missing files..."
Revert "Added some missing files that are required by other files and tests. These files were probably removed for a reason and their function might be changed and implemented in a different way but for now I will be adding them back since all the tests now pass and some things seem to be needed. We can later remove them again if need be."

This reverts commit daf64419d125663e543dd27324d1f2b350f88967.